// frontend2/frontend/types.ts
export interface Dentist {
  id: string;
  name: string;
  specialty: string;
  color: string;
}